klix/email_sender/requirements.txt
gspread
PyYAML
python-dotenv
tzdata
google-auth
oauthlib
requests
